feat: move generations export
