+++ packages/shared/src/server/__test__/index.ts
@@ -1,2 +1,3 @@
 export * from "./tracing-factory";
 export * from "./clickhouse-helpers";
+export * from "./org-factory";
