+++ packages/shared/src/server/__test__/org-factory.ts
@@ -1,0 +1,53 @@
+import { prisma } from "@langfuse/shared/src/db";
+import {
+  hashSecretKey,
+  getDisplaySecretKey,
+} from "@langfuse/shared/src/server";
+import { v4 } from "uuid";
+
+export function createBasicAuthHeader(
+  username: string,
+  password: string,
+): string {
+  const base64Credentials = Buffer.from(`${username}:${password}`).toString(
+    "base64",
+  );
+  return `Basic ${base64Credentials}`;
+}
+
+export type CreateOrgProjectAndApiKeyOptions = {
+  projectId?: string;
+};
+export const createOrgProjectAndApiKey = async (
+  props?: CreateOrgProjectAndApiKeyOptions,
+) => {
+  const projectId = props?.projectId ?? v4();
+  const org = await prisma.organization.create({
+    data: {
+      id: v4(),
+      name: v4(),
+    },
+  });
+  await prisma.project.create({
+    data: {
+      id: projectId,
+      name: v4(),
+      orgId: org.id,
+    },
+  });
+  const publicKey = v4();
+  const secretKey = v4();
+
+  const auth = createBasicAuthHeader(publicKey, secretKey);
+  await prisma.apiKey.create({
+    data: {
+      id: v4(),
+      projectId: projectId,
+      publicKey: publicKey,
+      hashedSecretKey: await hashSecretKey(secretKey),
+      displaySecretKey: getDisplaySecretKey(secretKey),
+    },
+  });
+
+  return { projectId, publicKey, secretKey, auth };
+};
