+++ web/src/__tests__/async/scores-api.servertest.ts
@@ -7,11 +7,9 @@ import {
   createObservationsCh,
   createScoresCh,
   createTracesCh,
-} from "@langfuse/shared/src/server";
-import {
   createOrgProjectAndApiKey,
-  makeZodVerifiedAPICall,
-} from "@/src/__tests__/test-utils";
+} from "@langfuse/shared/src/server";
+import { makeZodVerifiedAPICall } from "@/src/__tests__/test-utils";
 import { GetScoreResponse, GetScoresResponse } from "@langfuse/shared";
 import { prisma } from "@langfuse/shared/src/db";
 import { v4 } from "uuid";
