+++ web/src/__tests__/async/sessions-api.servertest.ts
@@ -1,10 +1,10 @@
-import {
-  createOrgProjectAndApiKey,
-  makeZodVerifiedAPICall,
-} from "@/src/__tests__/test-utils";
+import { makeZodVerifiedAPICall } from "@/src/__tests__/test-utils";
 import { v4 } from "uuid";
 import { prisma } from "@langfuse/shared/src/db";
-import { createTracesCh } from "@langfuse/shared/src/server";
+import {
+  createTracesCh,
+  createOrgProjectAndApiKey,
+} from "@langfuse/shared/src/server";
 import { createTrace } from "@langfuse/shared/src/server";
 import {
   GetSessionsV1Response,
